app/requirements.txt
aiofiles==0.4.0
Jinja2==2.10.3
python-dotenv==0.10.3
Pillow==6.2.0
